moved the git API to its own package, refactors
